Calculate step angle based on number of full steps
This removes the burden of determining the correct ratio for the dial
arc from the caller, which now only needs to pass in the step count for
a full circle, i.e., the step count of the outer dial.

Also, fix a bug in the angle calculation that was introduced with the
previous commit. To keep the math simpler, just wrap around to n+1
ticks unconditionally; for the full circle (i.e., outer dial) case,
the extra tick will just be obscured anyway, since it will overlap
with the first tick which is longer and thicker.